Add helper to determine zope.app.testing usability.
While z3c.testsetup can work also without zope.app.testing
installed, it runs into trouble when z.a.testing is installed but
not in a usable state. This is the case with Python 3.x at the
moment: you can install it, but it will be broken.

Therefore we need a way to determine clearly (in tests and code)
whether we can import *and use* zope.app.testing. The new function
is meant for that purpose.