Skip some tests if zope.app.testing not working.
With Python3 and missing working zope.app.testing we can (and should)
run other tests, but we can skip those tests that are clearly geared
towards accessibility of a working zope.app.testing. These are mainly
the functional tests.

Most tests still fail under Python3 and will have to be separated in
functional and not-functional tests.